Fixes: Slug on trick
