+++ composer.json
@@ -8,6 +8,7 @@
         "fzaninotto/faker": "^1.8",
         "sensio/framework-extra-bundle": "^5.1",
         "sensiolabs/security-checker": "^5.0",
+        "stof/doctrine-extensions-bundle": "^1.3",
         "symfony/apache-pack": "^1.0",
         "symfony/asset": "4.2.*",
         "symfony/console": "4.2.*",
