+++ templates/base.html.twig
@@ -14,6 +14,10 @@
         {% block seo %}{% endblock %}
     </head>
     <body>
+        {#LOADER AJAX#}
+        <div id="ajax_loader">
+            <i class="spinner-border text-primary" role="status"></i>
+        </div>
 
         <!-- HEADER OF PAGE -->
         <nav id="nav_desktop" style="display: none" class="align-items-center">
