+++ templates/macro/image.html.twig
@@ -23,10 +23,6 @@
 {% endmacro %}
 
 {% macro picker(form) %}
-    <div id="ajax_loader">
-        <i class="spinner-border text-primary" role="status"></i>
-    </div>
-
     <div id="media_picker" style="display: none;">
         <div id="media_content" class="row">
             <div class="col-12" style="height: 10%">
